JwtUtils and JwtRequestFilter classes added
The JwtUtils class is responsible for creating the JWT token, as well as for extracting the token fields contained in the requests received by the system. It is also responsible for calculating whether a given token has expired or not. IMPORTANT: Although it is not a relevant factor during development, the secret key should be placed in an environment file instead of in the class code.

The JwtRequestFilter class is responsible for preprocessing all requests received by the system that require authentication.